Fixed mutation error
You should not mutate the original objects as it could pose side effects. In this instance it was a string, so it is "safe", but it is a bad habit, so best to avoid where possible. Also in this instance it is nicer to use string templates rather than string concatenation.